requirements.txt
numpy==1.19.0
Pillow==8.1.1
grpcio==1.48.2
protobuf==3.17.1
